agrege ejercicios match

+++ 00-Unidades/03_match/Ejercicios_Match/Match_app_06.py
@@ -40,7 +40,7 @@ class App(customtkinter.CTk):
     
     def btn_informar_on_click(self):
         horas = self.txt_hora.get()
-        hora = float(horas)
+        hora = int(horas)
         
         match hora :
             case 7 | 8 | 9 | 10 | 11:
